Modify idle compile strategy as aot compiler crash
Issue: https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/IAQVWR

+++ ecmascript/platform/unix/linux/os.cpp
@@ -83,7 +83,7 @@ void BindMidCpuCore()
     cpu_set_t cpuset;
 
     CPU_ZERO(&cpuset);
-    for (int i = 0; i < 7; i++) { // 7: 0-3 little core, 4-6 mid core
+    for (size_t i = 0; i < 7; i++) { // 7: 0-3 little core, 4-6 mid core
         CPU_SET(i, &cpuset);
     }
